requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
